Add proper distrib command for linux and osx
This refactors to package up non-common libraries we use for osx and
linux. For linux in particular it's common to use sdl3 with sdl2-compat,
so for linux in particular we do detection to see if this is the setup,
and if so we bundle sdl3 as well. In this setup sdl2 dynamically points
at sdl3, so for the sdl2 library files to work in this setup, they need
to be bundled with sdl3.