Fix hash distribution problem leading to crash
Another engineer is working on a lua modding system, and when running
their mods they got an overflow in cache bucket zero which lead to a
panic. I wrote a test to reproduce the bug. It turns out that the
previous hashing function would not create an even distribution. The new
modding system produces text and graphics that seem to hash mostly to
bucket zero, causing an overflow.

This fixes that problem, by using a better hashing algorithm (FNV-1a)
for a more even distribution.

Includes tests that demonstrate that bucket zero essentially only has
sprite 0 and nothing else in it (the sprite for fog of war/black squares
for areas you shouldn't be able to see on the map), with a very even
distribution above it. Looks like our max depth is now 3, when we had
uneven buckets spilling in the hundreds before.

This should also drastically improve cache performance. With uneven
bucket distribution, the previous sytem was constrained by O(n)
pathalogical cases of walking long uneven cache lines.