Build updates for sdl3
Update pipeline, dockerfiles, and all local builds for sdl3 and
sdl3-mixer dependencies. Right now sdl3-mixer isn't officially
released (even though sdl3 has been out for over a year), but will be
release soon. Until then we are compiling sdl3 and sdl3-mixer from
source (so they are always version matched) with the exception that we
are vendoring for Windows, given that Windows is a bit of a trainwreck
to support building across msys2 shells, cmd.exe, powershell.exe, and
WSL2.