Use ideal gpu backends on each platform
On Windows in particular, opengl seems to very oddly set wrong brightness and gamma. Regardless, Direct3D is a much better choice on windows, so this ensures that it prefers this over opengl. In testing, this fixes the issue.

This also generally sets up the preferred backend for each
platform (Vulkan for Linux, Metal for OSX, and Direct3D for Windows).